pyproject.toml
[project]
name = "uniOAPI-MCP"
version = "0.1.0"
description = "Universal MCP server supporting any OpenAI-compatible LLM provider"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["mcp", "llm", "openai", "universal"]
dependencies = [
    "mcp[cli]>=1.13.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.0",
    "pyyaml>=6.0.2",
    "typer>=0.21.0",
]
